+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.19.4] - 2021-10-21
+- Fixed an issue where MonoScript bundle will attempt to pull in data it should not have.
+- Fixed an issue where NONRECURSIVE_DEPENDENCY_DATA would not properly calculate Shader Variants from Scenes.
+- Improved ArchiveAndCompress path handling to detect too long paths and log errors.
+- Improve performance of UpdateBundleObjectLayout build task.
+- Improved BuildCache to rebuild cache entries in case of exceptions instead of failing the build.
+
 ## [1.19.3] - 2021-09-14
 - Fixed an edge case where Link.xml ordering was not deterministic causing incremental player rebuilds to occur unnecessarily.
 - Fixed an issue where Cache Server integration was not update to use USerializer.